Add basic account management
This adds the django-user-accounts app, and configures it to provide
basic user account management functionality to mapit. Users can sign
up, login, reset passwords etc. I've added some very basic templates
copied from an example given by the django-user-accounts project and
then tweaked somewhat to fit into mapit's HTML.

There's still a bit of work to be done in styling this up so that it
looks and feels like the rest of mapit.