Upgrade to latest MapIt and Django 1.9
This updates this project to match the most recent changes in the MapIt app
and its example project, mostly for django 1.9 compatiblity and dropping
support for Django < 1.8.

Also changes the way this uses the existing project settings, by including
them from a separate file and then extending, rather than mixing MapIt
settings and overrides in the same file.