Ensure IPv6 addresses are handled by reset_ip_quotas
This adds separate loops to clear IPv6 address count and blocked keys to
the reset_ip_quotas management command.

Separate loops rather than a glob of the form `[.:]` are used to work
around an issue handling such globs in the mock used by our tests. Such
a glob would work in Redis itself, but this way we can be sure the tests
are covering the change.